requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.0.33
pandas==2.0.3
openpyxl==3.1.2
gunicorn==20.1.0